// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/customUserFields.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Definition of addition user registration fields
 */
export interface CustomUserFields { 
    /**
     * The field identifier.
     */
    key?: string;
    /**
     * The default field label.
     */
    label?: string;
    /**
     * The field is mandatory
     */
    required?: boolean;
    /**
     * List of authorized values for the field
     */
    values?: Array<string>;
}